refactor(report): remove Safe* type wrappers, use generated null/unknown guards
The codegen-framework now generates ValueFromObject methods with
null/unknown guards, making the hand-written Safe*Type wrappers
unnecessary. Extract the business-logic plan modifiers into their
own file and remove all wrapper plumbing.
